Rename resource.rs -> storage.rs
