+++ crates/dslab-storage/src/lib.rs
@@ -5,7 +5,7 @@ pub mod bandwidth;
 pub mod disk;
 pub mod events;
 pub mod fs;
-pub mod resource;
+pub mod storage;
 pub mod shared_disk;
 
 #[cfg(test)]
