+++ examples/master-workers/src/worker.rs
@@ -14,7 +14,7 @@ use dslab_network::model::*;
 use dslab_network::network::Network;
 use dslab_storage::disk::Disk;
 use dslab_storage::events::{DataReadCompleted, DataWriteCompleted};
-use dslab_storage::resource::Storage;
+use dslab_storage::storage::Storage;
 
 use crate::common::Start;
 use crate::task::*;
