+++ examples/storage-shared-disk-benchmark/src/main.rs
@@ -20,7 +20,7 @@ use dslab_core::simulation::Simulation;
 use dslab_core::{cast, log_error, log_info};
 
 use dslab_storage::events::{DataReadCompleted, DataReadFailed};
-use dslab_storage::resource::Storage;
+use dslab_storage::storage::Storage;
 use dslab_storage::shared_disk::SharedDisk;
 
 const SEED: u64 = 16;
